+++ source/extensions/omni.isaac.ros_bridge/python/tests/test_differential_base.py
@@ -27,7 +27,7 @@ from omni.isaac.core.utils.physics import simulate_async
 from omni.isaac.core_nodes.scripts.utils import set_target_prims
 from pxr import Gf, Sdf
 
-from .common import add_carter, add_carter_ros, set_rotate, set_translate, wait_for_rosmaster
+from .common import add_carter, add_carter_ros, set_rotate, set_translate, wait_for_rosmaster_async
 
 
 # Having a test class dervived from omni.kit.test.AsyncTestCase declared on the root of module will make it auto-discoverable by omni.kit.test
@@ -57,7 +57,7 @@ class TestRosDifferentialBase(omni.kit.test.AsyncTestCase):
         await omni.kit.app.get_app().next_update_async()
 
         self._roscore = Roscore()
-        await wait_for_rosmaster()
+        await wait_for_rosmaster_async()
         await omni.kit.app.get_app().next_update_async()
 
         try: